requirements.txt
aiogram==3.21.0
python-dotenv>=1.0.1
openai>=1.40.0
pdfplumber>=0.11.4
pypdf>=4.2.0
requests>=2.32.3
pymupdf>=1.24.0
pdf2image>=1.16.3
pillow>=10.0.0
